requirements.txt
evtx
openai-whisper
colorama
reportlab
psutil
